Owner can't edit or delete themself

+++ frontend/src/components/EditEmployees.jsx
@@ -1,20 +1,26 @@
-import { useState } from "react"
+import { useEffect, useState } from "react"
 import { useTranslation } from "react-i18next"
 import businessService from "../services/business"
+import userService from "../services/users"
 
 
-const ShowEmployee = ({ employee, handleEditEmployee, handleDeletion }) => {
+const ShowEmployee = ({ employee, handleEditEmployee, handleDeletion, currentUser }) => {
   console.log(employee)
+  
   return (
     <tr>
       <td>{employee[0]}</td> 
       <td>{employee[1]}</td>
-      <td>
-        <EditEmployeeForm employee={employee} handleEditEmployee={handleEditEmployee} />
-      </td>
-      <td>
-        <DeleteEmployeeForm employee={employee} handleDeletion={handleDeletion} />
-      </td>
+      {!(employee[0] === currentUser.Email) && (
+        <>
+        <td>
+          <EditEmployeeForm employee={employee} handleEditEmployee={handleEditEmployee} />
+        </td>
+        <td>
+          <DeleteEmployeeForm employee={employee} handleDeletion={handleDeletion} />
+        </td>
+        </>
+        )}
     </tr>
   )
 }
@@ -50,6 +56,15 @@ const DeleteEmployeeForm = ({ employee, handleDeletion }) => {
 
 const EditEmployees = ({ employees, onEmployeeEdited }) => {
   const { t, i18n } = useTranslation()
+  const [currentUser, setCurrentUser] = useState(null)
+
+  useEffect(() => {
+    const fetchUser = async () => {
+      const user = await userService.get()
+      setCurrentUser(user)
+    }
+    fetchUser()
+  }, [])
 
   const handleEditEmployee = async (event) => {
     event.preventDefault()
@@ -64,30 +79,31 @@ const EditEmployees = ({ employees, onEmployeeEdited }) => {
   const handleDeletion = async (event) => {
     event.preventDefault()
     console.log("Trying to delete employee");
+    
+    const currentUser = await userService.get()
+
     const formData = new FormData(event.target)
     const userEmail = formData.get("email")
     const businessId = (await businessService.get()).ID
 
+    if (userEmail === currentUser.Email) {
+      alert(t("alert.cannotdeleteyourself"))
+      return
+    }
+
+    const confirmed = window.confirm((t("alert.deletemember")) + userEmail + "?") 
+
+    if (!confirmed) {
+      return
+    }
+
     try {
       await businessService.deleteMembership(userEmail, businessId)
       console.log("Deleted:", userEmail);
-      console.log("Fetching updated employee list...");
       onEmployeeEdited()
     } catch (error) {
       console.error("Error deleting employee:", error);
     }
-
-    //console.log("Trying to delete employee:", employee);
-
-    // businessService.deleteMembership(employee)
-    //   .then(() => {
-    //     console.log("Fetching updated employee list...");
-    //     employeeGetter();
-    //   })
-    //   .catch(error => {
-    //     //täältä tulee vielä error, korjaantuu ma tai ti!
-    //     console.error("Error deleting employee:", error);
-    //   });
 	}
 
   return (
@@ -99,6 +115,7 @@ const EditEmployees = ({ employees, onEmployeeEdited }) => {
           employees.map(employee => (
             <ShowEmployee 
               employee={employee}
+              currentUser={currentUser}
               handleEditEmployee={handleEditEmployee}
               handleDeletion={handleDeletion}
               key={employee[0]}
